Combine support functions into a single module, utils
