--- multipcc/description.py
@@ -1,7 +1,0 @@
-def desciption(data):
-    description = data.root.description
-
-    description.current_simulation = 'Case 3: Directly calculate the radiative capture cross section by putting the photon density of states as 1/(2pi^2)((hbar omega)^2/(hbar c/eta_R)^3).'
-
-
-
